+++ test/components/index.js
@@ -1,0 +1,33 @@
+/*!
+ * Copyright (c) 2025 Digital Bazaar, Inc. All rights reserved.
+ */
+import * as brQuasar from '@bedrock/quasar';
+import * as brVue from '@bedrock/vue';
+import {createRouter, createWebHistory} from 'vue-router';
+import App from './App.vue';
+import iconSet from 'quasar/icon-set/fontawesome-v5.mjs';
+import {Quasar} from 'quasar';
+import '@bedrock/web-fontawesome';
+
+brVue.initialize({
+  async beforeMount({app}) {
+    // initialize br-quasar
+    await brQuasar.initialize({app});
+
+    // configure quasar
+    Quasar.iconSet.set(iconSet);
+
+    const router = createRouter({
+      history: createWebHistory(),
+      routes: [{
+        path: '/',
+        component: () => import('./HomePage.vue'),
+        meta: {title: 'Bedrock Vue Optical Scanner Test'}
+      }]
+    });
+    app.use(router);
+
+    // return root Vue component
+    return App;
+  }
+});