+++ test/dev.js
@@ -1,0 +1,10 @@
+/*!
+ * Copyright (c) 2025 Digital Bazaar, Inc. All rights reserved.
+ */
+import * as bedrock from '@bedrock/core';
+import '@bedrock/views';
+import '@bedrock/webpack';
+
+import './config.js';
+
+bedrock.start();