+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.0.0 - 2025-09-10
 
-### Changed
+### Added
 
-- Initial Changes
+- Initial release of Vue Optical Scanner components.
+- Integrates with @bedrock/web-opitcal-scanner for scanning logic.
+- Vue components for optical scanning UIs.